Move `FeedImageDataLoaderWithFallbackComposite` to production.
